chore: update test case
